backend/requirements.txt
# API Framework
fastapi>=0.95.0
uvicorn>=0.21.1
gunicorn>=20.1.0
pydantic>=1.10.7
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.9
alembic>=1.10.3
psycopg2-binary>=2.9.6

# Authentication
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
requests>=2.28.2
pydantic[email]>=1.10.7

# Testing
pytest>=7.3.1
httpx>=0.24.0
pytest-cov>=4.1.0
